webrtc/set-remote-description.html

<!DOCTYPE html>
<html>
<head>
  <title>RTCPeerConnection.setRemoteDescription exercises</title>
</head>
<body>
  <h2>SetRemoteDescription error handling</h2>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript">
    'use strict'

    promise_test(() => {
      const pc = new RTCPeerConnection(null);
      return pc.setRemoteDescription(
        {'type': 'offer', 'sdp': 'bogus'})
      .then(() => {
        assert_unreached('Bogus SDP should not be accepted');
      })
      .catch(e => {
        assert_equals(e.name, 'InvalidAccessError');
      });
    }, 'Bogus SDP');

    promise_test(() => {
      const pc = new RTCPeerConnection(null);
      return pc.setRemoteDescription(
        {'type': 'bogus', 'sdp': 'bogus'})
      .then(() => {
        assert_unreached('Bogus operation types should not be accepted');
      })
      .catch(e => {
        assert_equals(e.name, 'TypeError');
      });
    }, 'Bogus Operation');

    promise_test(() => {
      const pc = new RTCPeerConnection(null);
      return pc.setRemoteDescription(
        {'type': 'answer', 'sdp': 'bogus'})
      .then(() => {
        assert_unreached('Answer in "new" state should not be accepted');
      })
      .catch(e => {
        assert_equals(e.name, 'InvalidStateError');
      });
    }, 'Bogus operation for state');

  </script>
</body>
</html>
